refactor: Add new common/aws folder
